# backend/internal/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Ward API
  description: >
    Self-hosted security-edge control plane over Caddy + Coraza. Sessions and
    machine API tokens both authenticate as bearer tokens.
  version: "1"
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /waf-custom-rules:
    get:
      operationId: listWafCustomRules
      summary: List user-authored raw SecLang rules
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WafCustomRule" }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createWafCustomRule
      summary: Create a raw SecLang rule (validated against the edge before it's kept)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WafCustomRuleInput" }
      responses:
        "201":
          description: Created — the edge accepted the rule
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WafCustomRule" }
        default: { $ref: "#/components/responses/Error" }
  /waf-custom-rules/{id}:
    patch:
      operationId: updateWafCustomRule
      summary: Update a rule (partial; validated against the edge, reverted on rejection)
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      parameters: [{ $ref: "#/components/parameters/id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WafCustomRuleUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WafCustomRule" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteWafCustomRule
      summary: Delete a rule (reconciles the edge)
      parameters: [{ $ref: "#/components/parameters/id" }]
      responses:
        "204": { description: Deleted }
        default: { $ref: "#/components/responses/Error" }
...
  /settings:
    get:
      operationId: getSettings
      summary: Edge-wide settings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateSettings
      summary: Update whichever settings are present (audited; reconciles)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SettingsUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Settings" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Session token or machine API token.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    WafMode:
      type: string
      enum: [DetectionOnly, "On"]
...
    Settings:
      type: object
      required: [waf_engine_mode, acme_email, access_retention_days, waf_retention_days]
      properties:
        waf_engine_mode: { $ref: "#/components/schemas/WafMode" }
        acme_email: { type: string }
        access_retention_days: { type: integer }
        waf_retention_days: { type: integer }
        crs_version:
          type: string
          description: Read-only — OWASP CRS version the running edge reported.
        crowdsec_enabled: { type: boolean, description: Whether the CrowdSec bouncer is wired into the edge. }
        crowdsec_configured: { type: boolean, description: Read-only — LAPI URL + key are set in the deployment. }
        metrics_enabled: { type: boolean, description: Expose Prometheus metrics (per-host) at the internal admin /metrics. }
        log_level: { type: string, description: "Caddy's default logger level: DEBUG | INFO | WARN | ERROR." }
        access_log_errors_only: { type: boolean, description: Only write 5xx access-log entries (server errors). }
        tls_min_version: { type: string, description: "Edge-wide minimum TLS version: 1.2 | 1.3." }
        edge_versions:
          type: object
          additionalProperties: { type: string }
          description: Read-only — versions of the components compiled into the ward-caddy image this release targets (component → version). Ground truth is the image's OCI labels.
    SettingsUpdate:
      description: Only the fields present (non-zero) are applied.
      type: object
      properties:
        waf_engine_mode: { $ref: "#/components/schemas/WafMode" }
        acme_email: { type: string }
        access_retention_days: { type: integer, minimum: 1, maximum: 365 }
        waf_retention_days: { type: integer, minimum: 1, maximum: 365 }
        crowdsec_enabled: { type: boolean }
        metrics_enabled: { type: boolean }
        log_level: { type: string }
        access_log_errors_only: { type: boolean }
        tls_min_version: { type: string }
...
    Scope:
      type: string
      enum: [global, service]
...
    WafCustomRule:
      type: object
      required: [id, scope, name, seclang, enabled, created_at, updated_at]
      properties:
        id: { type: string }
        scope: { $ref: "#/components/schemas/Scope" }
        service_id: { type: string, nullable: true }
        name: { type: string }
        seclang: { type: string }
        enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WafCustomRuleInput:
      type: object
      required: [name, seclang]
      properties:
        scope: { $ref: "#/components/schemas/Scope" }
        service_id: { type: string, nullable: true }
        name: { type: string }
        seclang:
          type: string
          description: Raw SecLang (multi-line). Injected before the CRS include, after Ward's generated exclusions.
        enabled: { type: boolean }
    WafCustomRuleUpdate:
      description: >-
        JSON Merge Patch (RFC 7396): only the fields present in the body change. `null` clears a field to its default, nested objects merge field by field, arrays replace as a whole, and anything absent keeps its current value.
      type: object
      properties:
        scope: { $ref: "#/components/schemas/Scope" }
        service_id: { type: string, nullable: true }
        name: { type: string }
        seclang:
          type: string
          description: Raw SecLang (multi-line). Injected before the CRS include, after Ward's generated exclusions.
        enabled: { type: boolean }
...