# backend/internal/api/settings.go
package api

import (
	"encoding/json"
	"net/http"
	"net/mail"
	"strconv"
	"strings"

	"github.com/bhagyajitjagdev/ward/backend/internal/edge"
	"github.com/bhagyajitjagdev/ward/backend/internal/store"
)

type settingsDTO struct {
	WAFEngineMode       string `json:"waf_engine_mode"`                  // "DetectionOnly" | "On" — the global default
	ACMEEmail           string `json:"acme_email"`                       // contact email for managed (Let's Encrypt) certs
	AccessRetentionDays int    `json:"access_retention_days"`            // days of raw access events to keep
	WAFRetentionDays    int    `json:"waf_retention_days"`               // days of WAF detections to keep
	CRSVersion          string `json:"crs_version"`                      // read-only: OWASP CRS version the edge reported (from detections)
	CrowdSecEnabled     *bool  `json:"crowdsec_enabled,omitempty"`       // toggle the bouncer (pointer: distinguishes omitted from false on PATCH)
	CrowdSecConfigured  bool   `json:"crowdsec_configured"`              // read-only: LAPI URL + key present (env)
	MetricsEnabled      *bool  `json:"metrics_enabled,omitempty"`        // expose Prometheus metrics at the admin /metrics
	LogLevel            string `json:"log_level,omitempty"`              // Caddy default logger level: DEBUG|INFO|WARN|ERROR
	AccessLogErrorsOnly *bool  `json:"access_log_errors_only,omitempty"` // only log 5xx access entries
	TLSMinVersion       string `json:"tls_min_version,omitempty"`        // edge-wide TLS floor: "1.2" | "1.3"
	// EdgeVersions is read-only: the components compiled into the ward-caddy image this
	// release targets (component → version). Ground truth is the image's OCI labels.
	EdgeVersions map[string]string `json:"edge_versions,omitempty"`
}

// validWAFMode reports whether m is a valid engine mode. Empty is valid only for a
// per-service override, where it means "inherit the global default".
func validWAFMode(m string) bool {
	return m == "" || m == "DetectionOnly" || m == "On"
}

func (h *Handler) currentSettings(r *http.Request) settingsDTO {
	configured := h.crowdsec != nil
	csEnabled := h.store.CrowdSecEnabled(r.Context(), configured)
	metricsEnabled := h.store.MetricsEnabled(r.Context(), false)
	accessErrorsOnly := h.store.AccessLogErrorsOnly(r.Context(), false)
	return settingsDTO{
		WAFEngineMode:       h.store.WAFEngineMode(r.Context(), "DetectionOnly"),
		ACMEEmail:           h.store.ACMEEmail(r.Context(), ""),
		AccessRetentionDays: h.store.AccessRetentionDays(r.Context(), 7),
		WAFRetentionDays:    h.store.WAFRetentionDays(r.Context(), 30),
		CRSVersion:          h.store.LatestCRSVersion(r.Context()),
		CrowdSecEnabled:     &csEnabled,
		CrowdSecConfigured:  configured,
		MetricsEnabled:      &metricsEnabled,
		LogLevel:            h.store.LogLevel(r.Context(), "INFO"),
		AccessLogErrorsOnly: &accessErrorsOnly,
		TLSMinVersion:       h.store.TLSMinVersion(r.Context(), "1.2"),
		EdgeVersions:        edge.Versions(),
	}
}

func (h *Handler) getSettings(w http.ResponseWriter, r *http.Request) {
	writeJSON(w, http.StatusOK, h.currentSettings(r))
}

// updateSettings applies whichever fields are present (non-empty); each is
// validated and audited independently, then the config is reconciled once.
func (h *Handler) updateSettings(w http.ResponseWriter, r *http.Request) {
	var in settingsDTO
	if err := json.NewDecoder(r.Body).Decode(&in); err != nil {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "invalid JSON body"})
		return
	}
	changed := false
	if in.WAFEngineMode != "" {
		if in.WAFEngineMode != "DetectionOnly" && in.WAFEngineMode != "On" {
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "waf_engine_mode must be 'DetectionOnly' or 'On'"})
			return
		}
		if err := h.store.SetSetting(r.Context(), store.WAFModeKey, in.WAFEngineMode); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.WAFModeKey, in.WAFEngineMode)
		changed = true
	}
	if in.ACMEEmail != "" {
		if _, err := mail.ParseAddress(in.ACMEEmail); err != nil {
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "acme_email must be a valid email address"})
			return
		}
		if err := h.store.SetSetting(r.Context(), store.ACMEEmailKey, in.ACMEEmail); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.ACMEEmailKey, in.ACMEEmail)
		changed = true
	}
	if in.AccessRetentionDays > 0 {
		if in.AccessRetentionDays > 365 {
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "access_retention_days must be between 1 and 365"})
			return
		}
		if err := h.store.SetSetting(r.Context(), store.AccessRetentionKey, strconv.Itoa(in.AccessRetentionDays)); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.AccessRetentionKey, strconv.Itoa(in.AccessRetentionDays))
		changed = true
	}
	if in.WAFRetentionDays > 0 {
		if in.WAFRetentionDays > 365 {
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "waf_retention_days must be between 1 and 365"})
			return
		}
		if err := h.store.SetSetting(r.Context(), store.WAFRetentionKey, strconv.Itoa(in.WAFRetentionDays)); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.WAFRetentionKey, strconv.Itoa(in.WAFRetentionDays))
		changed = true
	}
	if in.CrowdSecEnabled != nil {
		if *in.CrowdSecEnabled && h.crowdsec == nil {
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "CrowdSec isn't configured — set WARD_CROWDSEC_API_URL and WARD_CROWDSEC_API_KEY in the deployment"})
			return
		}
		val := "0"
		if *in.CrowdSecEnabled {
			val = "1"
		}
		if err := h.store.SetSetting(r.Context(), store.CrowdSecEnabledKey, val); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.CrowdSecEnabledKey, val)
		changed = true
	}
	if in.MetricsEnabled != nil {
		val := "0"
		if *in.MetricsEnabled {
			val = "1"
		}
		if err := h.store.SetSetting(r.Context(), store.MetricsEnabledKey, val); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.MetricsEnabledKey, val)
		changed = true
	}
	if in.LogLevel != "" {
		lvl := strings.ToUpper(in.LogLevel)
		if lvl != "DEBUG" && lvl != "INFO" && lvl != "WARN" && lvl != "ERROR" {
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "log_level must be DEBUG, INFO, WARN or ERROR"})
			return
		}
		if err := h.store.SetSetting(r.Context(), store.LogLevelKey, lvl); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.LogLevelKey, lvl)
		changed = true
	}
	if in.AccessLogErrorsOnly != nil {
		val := "0"
		if *in.AccessLogErrorsOnly {
			val = "1"
		}
		if err := h.store.SetSetting(r.Context(), store.AccessLogErrorsKey, val); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.AccessLogErrorsKey, val)
		changed = true
	}
	if in.TLSMinVersion != "" {
		if in.TLSMinVersion != "1.2" && in.TLSMinVersion != "1.3" {
			writeJSON(w, http.StatusBadRequest, map[string]string{"error": "tls_min_version must be '1.2' or '1.3'"})
			return
		}
		if err := h.store.SetSetting(r.Context(), store.TLSMinVersionKey, in.TLSMinVersion); err != nil {
			writeErr(w, http.StatusInternalServerError, err)
			return
		}
		h.audit(r, "settings.update", store.TLSMinVersionKey, in.TLSMinVersion)
		changed = true
	}
	if !changed {
		writeJSON(w, http.StatusBadRequest, map[string]string{"error": "nothing to update"})
		return
	}
	// Regenerate: services inheriting the global WAF default flip; the ACME issuer picks up the email.
	h.reconcile(r.Context())
	writeJSON(w, http.StatusOK, h.currentSettings(r))
}